+++ CHANGELOG.md
@@ -3,10 +3,11 @@
 Note: 严格遵循[Semantic Versioning 2.0.0](http://semver.org/lang/zh-CN/)语义化版本规范。
 
 ## 4.1.0-rc
-`2017-10-17`
+`2017-10-19`
 
 * 【重做】`ChatView`，移除旧的API，减少数据冗余和流量消耗，并完美兼容数据缓存
 * 【新增】加入了`jquery`的依赖
+* 【优化】优化了`Search`和`SearchBar`
 
 ## 4.1.0-beta
 `2017-10-12`